fix: rating calculation to be consistent

+++ database/models/Review.ts
@@ -49,11 +49,11 @@ Review.init(
       },
     },
     rating: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.FLOAT,
       allowNull: false,
       validate: {
-        min: 1,
-        max: 5,
+        min: 1.0,
+        max: 5.0,
       },
     },
     title: {
